requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.1
databento==0.36.1
pandas==2.2.2
zstd==1.5.5.1